+++ astropy/coordinates/__init__.py
@@ -15,5 +15,6 @@ from .distances import *
 from .transformations import *
 from .builtin_systems import *
 from .name_resolve import *
+from .matching import *
 
 __doc__ += builtin_systems._transform_graph_docs