+++ CHANGES.rst
@@ -15,6 +15,15 @@ New Features
   - `astropy.coordinates.Distance` supports conversion to and from distance
     modulii.
 
+  - `astropy.coordinates.SphericalCoordinateBase` and derived classes now
+    support arrays of coordinates, enabling large speed-ups for some 
+    operations on multiple coordinates at the same time.
+
+  - Array coordinates can be matched to other array coordinates, finding the
+    closest matches between the two sets of coordiantes (see the 
+    `~astropy.coordinates.matching.match_coordinates_3d` and 
+    `~astropy.coordinates.matching.match_coordinates_sky` functions).
+
 - ``astropy.io.votable``
 
   - The format of the units of a VOTable file can be specified using